devtools/client/shared/webpack/shims/test/test_clipboard.html

<!DOCTYPE html>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1290230
-->
<head>
  <title>Test for Bug 1290230 - clipboard helpers</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script src="/tests/SimpleTest/EventUtils.js"></script>

<script type="application/javascript">
"use strict";
var exports = {}
</script>
</head>
<body onload="pre_do_tests()">
<script type="application/javascript">
"use strict";

const RESULT = "lark bunting";

function doCopy(e) {
  console.log(e.isTrusted);
  copyString(RESULT);
}

async function pre_do_tests() {
  // Temporarily allow content to access all resource:// URIs.
  await SpecialPowers.pushPrefEnv({
    set: [
      ["security.all_resource_uri_content_accessible", true]
    ]
  });

  // Load script.
  await (() => new Promise((resolve) => {
    var script = document.createElement("script");
    script.onload = resolve;
    script.src = "resource://devtools/client/shared/webpack/shims/platform-clipboard-stub.js";
    document.head.appendChild(script);
  }))();

  do_tests();
}

function do_tests() {
  let elt = document.querySelector("#key");
  elt.addEventListener("keydown", doCopy);

  // Set the clipboard to something other than what we expect.
  SpecialPowers.clipboardCopyString("snowy owl");

  elt.focus();
  sendString("x");

  is(SpecialPowers.getClipboardData("text/unicode"), RESULT, "clipboard copying worked");

  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();

</script>
<div id="key" tabindex="-1">Type Here</div>
</body>
